pyproject.toml
[project]
name = "webhook-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alpaca-py>=0.35.0",
    "fastapi>=0.115.6",
    "fastapi-utils>=0.8.0",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "autopep8>=2.3.1",
]
